clean up
- move all data into QueryRunnerTest
- next to create individual test queries with data/query/schema/table/stub
