+++ pinot-query-runtime/src/test/java/org/apache/pinot/query/runtime/queries/JoinQueryTest.java
@@ -1,0 +1,7 @@
+package org.apache.pinot.query.runtime.queries;
+
+import org.apache.pinot.query.runtime.QueryRunnerTestBase;
+
+
+public class JoinQueryTest extends QueryRunnerTestBase {
+}
