+++ pinot-query-runtime/src/test/java/org/apache/pinot/query/runtime/queries/SortQueryTest.java
@@ -1,0 +1,7 @@
+package org.apache.pinot.query.runtime.queries;
+
+import org.apache.pinot.query.runtime.QueryRunnerTestBase;
+
+
+public class SortQueryTest extends QueryRunnerTestBase {
+}
